fixed: reinterpret_cast

+++ libs/util/CAnimTask.cpp
@@ -590,7 +590,8 @@ void CAnimTask::progressUpdate(void* pVal) {
 void CAnimTask::postExecute(void* pResult, void* pExtraVal) {
 	__LOGT__( TAG, "postExecute()" );
 
-	int result = reinterpret_cast<int>(pResult);
+	//int result = reinterpret_cast<int>(pResult);
+	int result = static_cast<int>(reinterpret_cast<long>(pResult));
 	__LOGT__( TAG, "postExecute(): result = %s", (result? "TRUE" : "FALSE") );
 
 	//__LOGT__( TAG, "postExecute(): DONE = %s", get_done()? "TRUE" : "FALSE" );
