+++ libs/util/CProgressDlg.cpp
@@ -664,7 +664,8 @@ void CProgressDlg::progressUpdate(void* pVal) {
 void CProgressDlg::postExecute(void* pResult, void* pExtraVal) {
 	__LOGT__( TAG, "postExecute()" );
 
-	int result = reinterpret_cast<int>(pResult);
+	//int result = reinterpret_cast<int>(pResult);
+	int result = static_cast<int>(reinterpret_cast<long>(pResult));
 	__LOGT__( TAG, "postExecute(): result = %s", (result? "TRUE" : "FALSE") );
